move DrawerProvider to features
