+++ common/v2/features/Drawer/DrawerProvider.tsx
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
 import { Transition } from 'react-spring/renderprops.cjs';
 
-// @TODO Provider shouldn't depend on components.. refactor
-import { default as Overlay } from 'v2/components/Overlay';
-import { Drawer } from 'v2/features';
+import { Overlay } from 'v2/components';
+import { default as Drawer } from './Drawer';
 
 export interface Action {
   icon: string;
