+++ common/v2/providers/index.ts
@@ -1,4 +1,3 @@
-export { DrawerContext, default as DrawerProvider } from './DrawerProvider';
 export {
   NotificationsContext,
   NotificationsProvider,
